+++ pinyin_test.go
@@ -20,48 +20,48 @@ func testPinyin(s string, d map[*Args][][]string, f pinyinFunc) (t *testing.T) {
 
 func TestPinyin(t *testing.T) {
 	hans := "中国人"
-	test_data := map[*Args][][]string{
+	testData := map[*Args][][]string{
 		&Args{}: [][]string{
 			[]string{"zhong"},
 			[]string{"guo"},
 			[]string{"ren"},
 		},
-		&Args{Style: NORMAL}: [][]string{
+		&Args{Style: Normal}: [][]string{
 			[]string{"zhong"},
 			[]string{"guo"},
 			[]string{"ren"},
 		},
-		&Args{Style: TONE}: [][]string{
+		&Args{Style: Tone}: [][]string{
 			[]string{"zhōng"},
 			[]string{"guó"},
 			[]string{"rén"},
 		},
-		&Args{Style: TONE2}: [][]string{
+		&Args{Style: Tone2}: [][]string{
 			[]string{"zho1ng"},
 			[]string{"guo2"},
 			[]string{"re2n"},
 		},
-		&Args{Style: INITIALS}: [][]string{
+		&Args{Style: Initials}: [][]string{
 			[]string{"zh"},
 			[]string{"g"},
 			[]string{"r"},
 		},
-		&Args{Style: FIRST_LETTER}: [][]string{
+		&Args{Style: FirstLetter}: [][]string{
 			[]string{"z"},
 			[]string{"g"},
 			[]string{"r"},
 		},
-		&Args{Style: FINALS}: [][]string{
+		&Args{Style: Finals}: [][]string{
 			[]string{"ong"},
 			[]string{"uo"},
 			[]string{"en"},
 		},
-		&Args{Style: FINALS_TONE}: [][]string{
+		&Args{Style: FinalsTone}: [][]string{
 			[]string{"ōng"},
 			[]string{"uó"},
 			[]string{"én"},
 		},
-		&Args{Style: FINALS_TONE2}: [][]string{
+		&Args{Style: FinalsTone2}: [][]string{
 			[]string{"o1ng"},
 			[]string{"uo2"},
 			[]string{"e2n"},
@@ -73,7 +73,7 @@ func TestPinyin(t *testing.T) {
 		},
 	}
 
-	testPinyin(hans, test_data, Pinyin)
+	testPinyin(hans, testData, Pinyin)
 }
 
 func TestNoneHans(t *testing.T) {