+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Changelog
 
 ## 0.2.0 (yyyy-mm-dd)
+** 不兼容之前的版本 **
 
-* 新增 ``pinyin.NewArgs() *Args``
+* 新增 ``func NewArgs() *Args``
 * 各函数内的 ``a Args`` 参数改为 ``a *Args``
+* 规范命名:
+    * ``NORMAL`` -> ``Normal``
+    * ``TONE`` -> ``Tone``
+    * ``TONE2`` -> ``Tone2``
+    * ``INITIALS`` -> ``Initials``
+    * ``FIRST_LETTER`` -> ``FirstLetter``
+    * ``FINALS`` -> ``Finals``
+    * ``FINALS_TONE`` -> ``FinalsTone``
+    * ``FINALS_TONE2`` -> ``FinalsTone2``
 
 ## 0.1.1 (2014-12-07)
 * 更新拼音库